+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Added:
 - Added villager trade datapack system
+- Added breeding experience datapack system
+- Added fishing experience datapack system
 ### Fixed:
 - 
 ### Changed: